+++ src/main/java/model/com/iisquare/jees/oa/service/RoleService.java
@@ -84,7 +84,7 @@ public class RoleService extends ServiceBase {
 	}
 	
 	public int insert(Role persist) {
-		return roleDao.insert(persist);
+		return roleDao.insert(persist).intValue();
 	}
 	
 	public int update(Role persist) {
@@ -94,9 +94,9 @@ public class RoleService extends ServiceBase {
 	public int delete(Object... ids) {
 		String idStr = SqlUtil.buildSafeWhere(",", ids);
 		if(DPUtil.empty(idStr)) return 0;
-		int count = roleDao.getCount(DPUtil.stringConcat("parent_id in (", idStr, " )"), new Object[]{}, null);
+		int count = roleDao.getCount(DPUtil.stringConcat("parent_id in (", idStr, " )"), new Object[]{}, null).intValue();
 		if(count > 0) return -1;
-		count = memberDao.getCount(DPUtil.stringConcat("role_id in (", idStr, " )"), new Object[]{}, null);
+		count = memberDao.getCount(DPUtil.stringConcat("role_id in (", idStr, " )"), new Object[]{}, null).intValue();
 		if(count > 0) return -2;
 		return roleDao.deleteByIds(ids);
 	}
