+++ src/main/java/model/com/iisquare/jees/oa/service/SettingService.java
@@ -78,7 +78,7 @@ public class SettingService extends ServiceBase {
 		}
 		if(!DPUtil.empty(orderBy)) sb.append(" order by ").append(orderBy);
 		String sql = sb.toString();
-		int total = settingDao.getCount(sql, paramMap, true);
+		int total = settingDao.getCount(sql, paramMap, true).intValue();
 		sql = DPUtil.stringConcat(sql, SqlUtil.buildLimit(page, pageSize));
 		List<Map<String, Object>> rows = settingDao.npJdbcTemplate().queryForList(sql, paramMap);
 		rows = ServiceUtil.fillRelations(rows, memberDao, new String[]{"operate_id"}, new String[]{"serial", "name"}, null);
@@ -121,7 +121,7 @@ public class SettingService extends ServiceBase {
 	}
 	
 	public int insert(Setting persist) {
-		return settingDao.insert(persist);
+		return settingDao.insert(persist).intValue();
 	}
 	
 	public int update(Setting persist) {
